[QA] Generate Code Owners
Record to codeowners.

Types fixup

Not done yet, but so far I've got a sorta-solution for
just the files, not the excludeFiles.

The rub is that one of my expressions breaks when I use the real
rule set.  There's something in the real rule set that I'm not
considering.  But, I'm sleepy lol

Meh, for codeowners, this is done.